@@ -101,6 +64,57 @@
     "feathers-socketio": "^1.4.1",
     "mongoose": "^4.6.3",
     "rxjs": "~5.0.2",
-    "zone.js": "~0.7.4"
+    "zone.js": "~0.7.6"
+  },
+  "devDependencies": {
+    "@angular/compiler-cli": "2.4.8",
+    "@angularclass/hmr": "1.2.2",
+    "@angularclass/hmr-loader": "3.0.2",
+    "@ngrx/store-devtools": "3.2.2",
+    "@ngrx/store-log-monitor": "3.0.2",
+    "@types/body-parser": "^0.0.34",
+    "@types/bunyan": "^0.0.35",
+    "@types/chalk": "^0.4.31",
+    "@types/compression": "^0.0.33",
+    "@types/core-js": "^0.9.35",
+    "@types/cors": "0.0.31",
+    "@types/mongoose": "^4.5.38",
+    "@types/node": "^7.0.0",
+    "add-asset-html-webpack-plugin": "^1.0.2",
+    "angular2-template-loader": "^0.6.0",
+    "assets-webpack-plugin": "^3.5.1",
+    "awesome-typescript-loader": "~3.0.0-beta.18",
+    "codelyzer": "~2.0.0-beta.4",
+    "copy-webpack-plugin": "^4.0.0",
+    "css-loader": "^0.26.0",
+    "exports-loader": "^0.6.3",
+    "expose-loader": "^0.7.1",
+    "extract-text-webpack-plugin": "~2.0.0-rc.3",
+    "file-loader": "^0.10.0",
+    "html-webpack-plugin": "^2.28.0",
+    "json-loader": "^0.5.4",
+    "ng-router-loader": "^2.1.0",
+    "ngc-webpack": "1.1.0",
+    "node-sass": "^4.5.0",
+    "nodemon": "^1.11.0",
+    "optimize-js-plugin": "0.0.4",
+    "raw-loader": "^0.5.1",
+    "rimraf": "^2.5.2",
+    "sass-loader": "^6.0.0",
+    "script-ext-html-webpack-plugin": "^1.6.2",
+    "style-loader": "^0.13.1",
+    "to-string-loader": "^1.1.4",
+    "ts-node": "^2.0.0",
+    "tslib": "^1.5.0",
+    "tslint": "~4.4.2",
+    "tslint-loader": "^3.3.0",
+    "typescript": "~2.1.6",
+    "url-loader": "^0.5.7",
+    "webpack": "2.2.0",
+    "webpack-dev-middleware": "^1.10.0",
+    "webpack-dev-server": "2.4.1",
+    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
+    "webpack-hot-middleware": "2.17.1",
+    "webpack-merge": "~3.0.0"
   }
 }
