@@ -2,16 +2,22 @@
  * @author: hjl
  */
 
-const webpack = require('webpack')
-import helpers from './helpers'
+import { helpers } from './helpers'
 const webpackMerge = require('webpack-merge') // used to merge webpack configs
-import commonConfig from './webpack.common' // the settings that are common to prod and dev
+const webpackMergeDll = webpackMerge.strategy({ plugins: 'replace' })
+import { webpackCommonConfig } from './webpack.common' // the settings that are common to prod and dev
 
 /**
  * Webpack Plugins
  */
 const DefinePlugin = require('webpack/lib/DefinePlugin')
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin')
+const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin')
+const NamedModulesPlugin = require('webpack/lib/NamedModulesPlugin')
+const DllBundlesPlugin = require('webpack-dll-bundles-plugin').DllBundlesPlugin
+
+
+
 
 
 /**
@@ -29,12 +35,15 @@ const METADATA = webpackMerge({}, {
   HMR: HMR
 })
 
+
+
 /**
  * Webpack configuration
  *
  * See: http://webpack.github.io/docs/configuration.html#cli
  */
-export default webpackMerge(commonConfig({ env: ENV }), {
+
+export default webpackMerge(webpackCommonConfig({ env: ENV }), {
 
 
   /**
@@ -93,7 +102,7 @@ export default webpackMerge(commonConfig({ env: ENV }), {
      *
      * See: http://webpack.github.io/docs/configuration.html#output-sourcemapfilename
      */
-    sourceMapFilename: '[name].bundle.map',
+    sourceMapFilename: '[file].map',
 
     /** The filename of non-entry chunks as relative path
      * inside the output.path directory.
@@ -128,6 +137,52 @@ export default webpackMerge(commonConfig({ env: ENV }), {
       }
     }),
 
+    new DllBundlesPlugin({
+      bundles: {
+        polyfills: [
+          'core-js',
+          {
+            name: 'zone.js',
+            path: 'zone.js/dist/zone.js'
+          },
+          {
+            name: 'zone.js',
+            path: 'zone.js/dist/long-stack-trace-zone.js'
+          },
+        ],
+        vendor: [
+          '@angular/platform-browser',
+          '@angular/platform-browser-dynamic',
+          '@angular/core',
+          '@angular/common',
+          '@angular/forms',
+          '@angular/http',
+          '@angular/router',
+          '@angularclass/hmr',
+          'rxjs',
+        ]
+      },
+      dllDir: helpers.root('dll'),
+      webpackConfig: webpackMergeDll(webpackCommonConfig({ env: ENV }), {
+        devtool: 'cheap-module-source-map',
+        plugins: []
+      })
+    }),
+
+    /**
+     * Plugin: AddAssetHtmlPlugin
+     * Description: Adds the given JS or CSS file to the files
+     * Webpack knows about, and put it into the list of assets
+     * html-webpack-plugin injects into the generated html.
+     *
+     * See: https://github.com/SimenB/add-asset-html-webpack-plugin
+     */
+    new AddAssetHtmlPlugin([
+      { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('polyfills')}`) },
+      { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('vendor')}`) }
+    ]),
+
+
     /**
      * Plugin LoaderOptionsPlugin (experimental)
      *
@@ -136,26 +191,13 @@ export default webpackMerge(commonConfig({ env: ENV }), {
     new LoaderOptionsPlugin({
       debug: true,
       options: {
-
-        /**
-         * Static analysis linter for TypeScript advanced options configuration
-         * Description: An extensible linter for the TypeScript language.
-         *
-         * See: https://github.com/wbuchwalter/tslint-loader
-         */
-        // tslint: {
-        //   emitErrors: false,
-        //   failOnHint: false,
-        //   resourcePath: 'client'
-        // },
-
       }
     }),
 
 
-    new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoErrorsPlugin()
+    // new webpack.optimize.OccurrenceOrderPlugin(),
+    // new webpack.HotModuleReplacementPlugin(),
+    // new webpack.NoErrorsPlugin()
   ],
 
   /**
@@ -191,5 +233,4 @@ export default webpackMerge(commonConfig({ env: ENV }), {
     clearImmediate: false,
     setImmediate: false
   }
-
 })
