@@ -3,7 +3,7 @@
  */
 
 const webpack = require('webpack')
-import helpers from '../../client/config/helpers'
+import { helpers } from '../../client/config/helpers'
 
 /*
  * Webpack Plugins
@@ -27,8 +27,8 @@ const fs = require('fs')
 // Look here http://jlongster.com/Backend-Apps-with-Webpack--Part-I
 const nodeModules = {}
 fs.readdirSync('node_modules')
-    .filter(x => ['.bin'].indexOf(x) === -1)
-    .forEach(mod => nodeModules[mod] = 'commonjs ' + mod)
+  .filter(x => ['.bin'].indexOf(x) === -1)
+  .forEach(mod => nodeModules[mod] = 'commonjs ' + mod)
 
 /*
  * Webpack configuration
@@ -37,22 +37,6 @@ fs.readdirSync('node_modules')
  */
 export default {
   target: 'node',
-
-  /*
-   * Static metadata for index.html
-   *
-   * See: (custom attribute)
-   */
-  metadata: METADATA,
-
-  /*
-   * Cache generated modules and chunks to improve performance for multiple incremental builds.
-   * This is enabled by default in watch mode.
-   * You can pass false to disable it.
-   *
-   * See: http://webpack.github.io/docs/configuration.html#cache
-   */
-  cache: false,
 
   /*
    * The entry point for the bundle
@@ -68,7 +52,7 @@ export default {
     libraryTarget: 'commonjs2'
   },
 
-  devtool: 'source-map',
+  devtool: 'eval',
   externals: nodeModules,
 
 
@@ -84,14 +68,8 @@ export default {
      *
      * See: http://webpack.github.io/docs/configuration.html#resolve-extensions
      */
-    extensions: ['', '.ts', '.js', '.json'],
-
-    // Make sure client root is client
-    root: helpers.root('app'),
-
-    // remove other default values
-    modulesDirectories: ['node_modules'],
-
+    extensions: ['.ts', '.js', '.json'],
+    modules: [helpers.root('app'), helpers.root('node_modules')],
   },
 
   /*
@@ -101,19 +79,6 @@ export default {
    */
   module: {
     /*
-     * An array of applied pre and post loaders.
-     *
-     * See: http://webpack.github.io/docs/configuration.html#module-preloaders-module-postloaders
-     */
-    preLoaders: [
-      {
-        test: /\.(ts|js)$/,
-        loader: 'tslint',
-        exclude: /node_modules|bower_components|webpack/,
-      },
-    ],
-
-    /*
      * An array of automatically applied loaders.
      *
      * IMPORTANT: The loaders here are resolved relative to the resource which they are applied to.
@@ -121,21 +86,17 @@ export default {
      *
      * See: http://webpack.github.io/docs/configuration.html#module-loaders
      */
-    loaders: [
-      /*
-      * Typescript loader support for .ts and Angular 2 async routes via .async.ts
-      * Replace templateUrl and stylesUrl with require()
-      *
-      * See: https://github.com/s-panferov/awesome-typescript-loader
-      * See: https://github.com/TheLarkInn/angular2-template-loader
-      */
+    rules: [
       {
-        test: /\.(ts|js)$/,
-        loaders: [
-          'awesome-typescript-loader',
+        test: /\.ts$/,
+        use: [
+          {
+            loader: 'awesome-typescript-loader',
+            options: {
+              configFileName: 'tsconfig.json'
+            }
+          },
         ],
-        query: { forkChecker: true },
-        // exclude: [/\.(spec|e2e)\.ts$/, /node_modules/]
         exclude: [/node_modules/]
       },
 
@@ -146,45 +107,11 @@ export default {
        */
       {
         test: /\.json$/,
-        loader: 'json-loader'
+        use: 'json-loader'
       },
     ]
 
   },
-
-  /*
-   * Add additional plugins to the compiler.
-   *
-   * See: http://webpack.github.io/docs/configuration.html#plugins
-   */
-  plugins: [
-
-    /*
-     * Plugin: ForkCheckerPlugin
-     * Description: Do type checking in a separate process, so webpack don't need to wait.
-     *
-     * See: https://github.com/s-panferov/awesome-typescript-loader#forkchecker-boolean-defaultfalse
-     */
-    new ForkCheckerPlugin(),
-
-    /**
-     * Plugin: DefinePlugin
-     * Description: Define free variables.
-     * Useful for having development builds with debug logging or adding global constants.
-     *
-     * Environment helpers
-     *
-     * See: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
-     */
-    // NOTE: when adding more properties, make sure you include them in custom-typings.d.ts
-    // new DefinePlugin({
-    //   'ENV': JSON.stringify(METADATA.ENV),
-    //   'process.env': {
-    //     'ENV': JSON.stringify(METADATA.ENV),
-    //     'NODE_ENV': JSON.stringify(METADATA.ENV)
-    //   }
-    // }),
-  ],
 
   /*
    * Include polyfills or mocks for various node stuff
