@@ -1,2 +1,1 @@
-import { ModModule } from './mod.module'
-export default ModModule
+export { ModModule } from './mod.module'
