@@ -1,4 +1,4 @@
-export default {
+export const commonConfig = {
     port: 80,
     tokenKey: 'famn-jwt'
 }
