@@ -1,5 +1,5 @@
-import commonConfig from './config.common'
+import { commonConfig } from './config.common'
 
-export default Object.assign({}, commonConfig, {
-    port: '3030'
+export const devConfig = Object.assign({}, commonConfig, {
+  port: '3030'
 })
